Create a simple blog websiteO
A new project foodsite and its first app foodieblog  created for a website to allow posts and comments on different
foods. Postgresql Database used to store the data from site. Models added to map to tables in the database.

Models registered to the admin inbuilt panel. Admin classes created to customize the admin interface.

Views and templates added to generate dynamic and the static HTML output, to perform logic and return a response to user.